Add a CLI utility to generate html from .cshtml templates
The CLI is splitted into two projects: a console application and a RCL that holds all cshtml templates.
The reason is that the engine requires to have cshtml file inside a razor class library to work.
