+++ src/HtmlGenerator/Smart.Design.HtmlGenerator/CommandOptions/ListingOptions.cs
@@ -2,10 +2,10 @@ using CommandLine;
 
 namespace Smart.Design.HtmlGenerator.CommandOptions
 {
-    [Verb("list", HelpText = "Lists design elements")]
+    [Verb("list", HelpText = "Lists design components")]
     public class ListingOptions
     {
-        [Option('c', "category", Required = false, HelpText = "Category of design elements")]
+        [Option('c', "category", Required = false, HelpText = "Category of design components")]
         public string DesignElementCategory { get; set; }
     }
 }
