+++ src/HtmlGenerator/Smart.Design.HtmlGenerator/Managers/IComponentDiscoverer.cs
