+++ src/HtmlGenerator/Smart.Design.HtmlGenerator/Models/CommandResponse.cs
@@ -1,6 +1,3 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -20,8 +17,9 @@ namespace Smart.Design.HtmlGenerator.Models
         {
             if (IsSuccess) return SuccessMessage;
 
-            StringBuilder stringBuilder = new("Exception Stacks");
-            Errors.ForEach(error => stringBuilder.AppendLine((string?) $"exception with message => {error}"));
+            StringBuilder stringBuilder = new("Exception Stack");
+            stringBuilder.AppendLine();
+            Errors.ForEach(error => stringBuilder.AppendLine((string?) $"Exception with message => {error}"));
             return stringBuilder.ToString();
         }
