+++ src/HtmlGenerator/Smart.Design.HtmlGenerator/Models/Enumerations/CommandType.cs
