+++ src/HtmlGenerator/Smart.Design.HtmlGenerator/Models/Exceptions/CustomExceptions.cs
@@ -1,6 +1,3 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
 using System;
 
 namespace Smart.Design.HtmlGenerator.Models.Exceptions
