+++ src/HtmlGenerator/Smart.Design.HtmlGenerator/Properties/launchSettings.json
