+++ src/HtmlGenerator/Smart.Design.HtmlGenerator/nlog.config
