+++ src/HtmlGenerator/Smart.Design.Templates/Smart.Design.Templates.csproj
