+++ src/HtmlGenerator/Smart.Design.Templates/Views/application-form.cshtml
